Feat : OAuth + Session , JWT

+++ pages/api/delete.js
@@ -1,5 +1,7 @@
 import { connectDB } from "@/util/database";
 import { ObjectId } from "mongodb";
+import { getServerSession } from "next-auth";
+import { authOptions } from "./auth/[...nextauth]";
 
 export default async function Delete(요청,응답){
 
@@ -7,16 +9,28 @@ export default async function Delete(요청,응답){
         return res.status(405).json({ error: "허용되지 않은 요청 방식입니다." });
     }
     
+    let session = await getServerSession(요청,응답,authOptions)
+
     try{
         //JSON : Object형태에서 속성이 큰 따옴표로 되어 있는 것
         //JSON -> Object변환은 JSON.parse()를 쓰자
         const client = await connectDB;
         const db = await client.db("forum");
-        console.log(요청.body);
-        let id = 요청.body;
 
-        let result = await db.collection('post').deleteOne({_id:new ObjectId(id)})
-        console.log(result)
+        let id = 요청.body;
+        let target = await db.collection('post').findOne({_id: new ObjectId(id)})
+        let result
+
+
+        if(!session){
+            return 응답.status(500).json({error:"로그인 필요"})
+        }
+        else{
+           if(session.user.email == target.author || session.user.role == 'admin'){
+                result = await db.collection('post').deleteOne({_id:new ObjectId(id)});
+            }
+        }
+        
         
         switch(result.deletedCount){
             case 0:
